create new notification watcher
